ref(samples): Move KafkaShowcase to kafka subpackage
Move KafkaShowcase under io.sentry.samples.console.kafka and update Main to
import the relocated class.

This keeps Kafka-specific sample code grouped in a dedicated package
without changing runtime behavior.
